module 'typedefs' moved to package 'common'. server types moved to distinct module.
