+++ pjrpc/client/__init__.py
@@ -2,13 +2,15 @@
 JSON-RPC client.
 """
 
-from .client import AbstractAsyncClient, AbstractClient, AsyncBatch, Batch
+from .client import AbstractAsyncClient, AbstractClient, AsyncBatch, BaseAbstractClient, BaseBatch, Batch
 from .tracer import LoggingTracer, Tracer
 
 __all__ = [
     'AbstractClient',
     'AbstractAsyncClient',
     'AsyncBatch',
+    'BaseBatch',
+    'BaseAbstractClient',
     'Batch',
     'LoggingTracer',
     'Tracer',
